* mcc: applied patch kindly provided by Alfonso Ranieri which will make   our sources compile/run on MorphOS and bring them to a more compatible   level so that hopefully we will soon have a completly running version   for all platforms. However, there is still heaps of work to do to   stabilize the sources - especially for different GCC versions -   but we are getting near. Thanks again, Alfonso.

+++ mcc/SharedData.cpp
@@ -69,7 +69,7 @@ VOID SharedData::Setup (struct Screen *scr)
 
   Scr = scr;
   InitConfig();
-  DoMethod(_win(Obj), MUIM_Window_AddEventHandler, &Events);
+  DoMethod(_win(Obj), MUIM_Window_AddEventHandler, (ULONG)&Events);
 
   LEAVE();
 }
@@ -78,7 +78,7 @@ VOID SharedData::Cleanup ()
 {
   ENTER();
 
-  DoMethod(_win(Obj), MUIM_Window_RemEventHandler, &Events);
+  DoMethod(_win(Obj), MUIM_Window_RemEventHandler, (ULONG)&Events);
   FreeConfig();
 
   AnimTimer.Stop(Obj);
